refactor: consolidate models.py, models_v2.py, models_state.py into single file (#291)
All 15 model classes now in bot/database/models.py.
Deleted models_v2.py and models_state.py.
Updated all imports across codebase (11 files).
